Moved parsing files to src/compiler/frontend
