+++ src/main.c
@@ -10,7 +10,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "parser.h"
+#include "compiler/frontend/parser.h"
 
 int main(int argc, char** argv)
 {
